Reestructurar directorios
Reestructuracion de directorios y archivos siguiendo este formato/ejemplo
/La-Cosa-API
        README.md
        requeriments.txt (hay que hacerlo)
        /src
                main.py
                settings.py
                version.py(lo tiene lean no se si ser necesario)
                constants.py (de ser necesarias)
                api.py (define la API incluyendo todos los routers)
                /models
                        db.py (instancia una Database de pony)
                /games (todo lo relacionado a partidas)
                        models.py (su modelo en DB)
                        schemas.py (sus esquemas de representacion por casos)
                        utils.py (De necesitar alguna funcion extra para
                                  manipular games en la api)
                        crud.py (archivo con funciones para Create, Read, Update, Delete
                                 de esos modelos)
                        endpoints.py (un router con los endpoints para games)
                /players
                        idem
                /cards
                        idem
        /tests
                /games_tests
                        populate_database.py
                        game_tests.py
                        etc...
