--- main.py
@@ -1,9 +1,0 @@
-from fastapi import FastAPI
-
-app = FastAPI()
-
-
-@app.get("/")
-async def root():
-    return {"message": "La Cosa"}
-
