+++ src/models/db.py
@@ -1,0 +1,3 @@
+from pony.orm import Database
+
+db = Database()
