+++ src/settings.py
@@ -2,4 +2,4 @@ import os
 
 ENVIRONMENT = os.getenv('LaCosaEnv', 'test')
 
-DATABASE_FILENAME = f"database_{ENVIRONMENT}.sqlite"
+DATABASE_FILENAME = f"database_la_cosa.sqlite"
